test: share publisher html fixture walker
